refactoring model modules
