+++ cell2location/models/base/base_model.py
