+++ cell2location/models/base/pymc3_model.py
@@ -9,7 +9,7 @@ import theano
 from pymc3.variational.callbacks import CheckParametersConvergence
 from tqdm.auto import tqdm
 
-from cell2location.models.base_model import BaseModel
+from cell2location.models.base.base_model import BaseModel
 
 
 # base model class - defining shared methods but not the model itself
