+++ cell2location/models/simplified/LocationModelHierarchicalWMultiExperiment.py
@@ -1,19 +1,54 @@
 # -*- coding: utf-8 -*-
-r"""Location model decomposes the expression of genes across locations into a set of reference regulatory programmes,
-    it is identical to LocationModelLinearDependentW but does not account for correlation of programs
-    across locations with similar cell composition, thus has reduced accuracy."""
+r"""The Co-Location model decomposes the expression of genes across locations measured in multiple experiments
+    into a set of reference regulatory programmes, while accounting for correlation of programs
+    across locations with similar cell composition.
+    Overdispersion alpha_eg & additive background s_eg for each experiment and gene."""
 
-import matplotlib.pyplot as plt
 import numpy as np
+import pandas as pd
 import pymc3 as pm
 import theano.tensor as tt
+import theano
 
-from cell2location.models.pymc3_loc_model import Pymc3LocModel
+from cell2location.models.base.pymc3_loc_model import Pymc3LocModel
 
 
 # defining the model itself
-class LocationModelHierarchicalW(Pymc3LocModel):
-    r"""Provided here as a 'base' model.
+class LocationModelHierarchicalWMultiExperiment(Pymc3LocModel):
+    r"""Cell2location models the elements of :math:`D` as Negative Binomial distributed,
+    given an unobserved rate :math:`mu` and a gene-specific over-dispersion parameter :math:`\alpha_g`
+    which describes variance in expression of individual genes that is not explained by the regulatory programs:
+    
+    .. math::
+        D_{s,g} \sim \mathtt{NB}(\mu_{s,g}, \alpha_{e,g})
+    
+    The containment prior on overdispersion :math:`\alpha_{e,g}` parameter is used
+    (for more details see: https://statmodeling.stat.columbia.edu/2018/04/03/justify-my-love/).
+    
+    The spatial expression levels of genes :math:`\mu_{s,g}` in the rate space are modelled
+    as the sum of five non-negative components:
+    
+    .. math::
+        \mu_{s,g} = m_{g} \left (\sum_{f} {w_{s,f} \: g_{f,g}} \right) + l_s + s_{e,g}
+    
+    Here, :math:`w_{s,f}` denotes regression weight of each program :math:`f` at location :math:`s` ;
+    :math:`g_{f,g}` denotes the reference signatures of cell types :math:`f` of each gene :math:`g` - input to the model;
+    :math:`m_{g}` denotes a gene-specific scaling parameter which accounts for difference
+    in the global expression estimates between technologies;
+    :math:`l_{s}` and :math:`s_{e,g}` are additive components that capture additive background variation
+    that is not explained by the bi-variate decomposition.
+    
+    The prior distribution on :math:`w_{s,f}` is chosen to reflect the absolute scale. This is done by inferring a hierarchical prior representing per location number of cell types and cell density.
+    
+    This prior is specified using 3 `cell_number_prior` input parameters:
+    
+    * **cells_per_spot** is derived from examining the paired histology image to get an idea about
+      the average nuclei count per location.
+    
+    * **factors_per_spot** reflects the number of regulatory programmes / cell types you expect to find in each location.
+    
+    The prior distribution on :math:`m_{g}` is informed by the expected change in sensitivity from single cell to spatial
+    technology, and is specified in `gene_level_prior`.
 
     Parameters
     ----------
@@ -21,32 +56,59 @@ class LocationModelHierarchicalW(Pymc3LocModel):
         Pandas data frame with gene programmes - genes in rows, cell types / factors in columns
     X_data :
         Numpy array of gene expression (cols) in spatial locations (rows)
+    n_comb :
+        The number of co-located cell type combinations (in the prior).
+        The model is fairly robust to this choice when the prior has low effect on location weights W
+        (`spot_fact_mean_var_ratio` parameter is low), but please use the default unless know what you are doing (Default: 50)
     n_iter :
         number of training iterations
     learning_rate, data_type, total_grad_norm_constraint, ...:
         See parent class BaseModel for details.
     gene_level_prior :
-        see the description for CoLocationModelNB4V2
+        prior on change in sensitivity between single cell and spatial technology (**mean**),
+        how much individual genes deviate from that (**sd**),
+        
+        * **mean** a good choice of this prior for 10X Visium data and 10X Chromium reference is between 1/3 and 1 depending
+          on how well each experiment worked. A good choice for SmartSeq 2 reference is around ~ 1/10.
+        * **sd** a good choice of this prior is **mean** / 2.
+          Avoid setting **sd** >= **mean** because it puts a lot of weight on 0.
     gene_level_var_prior :
-        see the description for CoLocationModelNB4V2
+        Certainty in the gene_level_prior (mean_var_ratio)
+        - by default the variance in our prior of mean and sd is equal to the mean and sd
+        decreasing this number means having higher uncertainty in the prior
     cell_number_prior :
-        see the description for CoLocationModelNB4V2, this model does not have **combs_per_spot**
-        parameter.
+        prior on cell density parameter:
+        
+        * **cells_per_spot** - what is the average number of cells you expect per location? This could also be the nuclei
+          count from the paired histology image segmentation.
+        * **factors_per_spot** - what is the number of cell types
+          number of factors expressed per location?
     cell_number_var_prior :
-        see the description for CoLocationModelNB4V2, this model does not have
-        **combs_mean_var_ratio** parameter.
+        Certainty in the cell_number_prior (cells_mean_var_ratio, factors_mean_var_ratio)
+        - by default the variance in the value of this prior is equal to the value of this itself.
+        decreasing this number means having higher uncertainty in the prior
     phi_hyp_prior :
-        see the description for CoLocationModelNB4V2
+        prior on NB alpha overdispersion parameter, the rate of exponential distribution over alpha.
+        This is a containment prior so low values mean low deviation from the mean of NB distribution.
+        
+        * **mu** average prior
+        * **sd** standard deviation in this prior
+        When using the Visium data model is not sensitive to the choice of this prior so it is better to use the default.
+    spot_fact_mean_var_ratio :
+        the parameter that controls the strength of co-located cell combination prior on
+        :math:`w_{s,f}` density across locations. It is expressed as mean / variance ratio with low values corresponding to
+        a weakly informative prior. Use the default value of 0.5 unless you know what you are doing.
 
     Returns
     -------
 
     """
-# LocationModelNB4V7_V4_V4
+# CoLocationModelNB4E6V2
     def __init__(
             self,
             cell_state_mat: np.ndarray,
             X_data: np.ndarray,
+            n_comb: int = 50,
             data_type: str = 'float32',
             n_iter=20000,
             learning_rate=0.005,
@@ -60,8 +122,11 @@ class LocationModelHierarchicalW(Pymc3LocModel):
                                'factors_per_spot': 7},
             cell_number_var_prior={'cells_mean_var_ratio': 1,
                                    'factors_mean_var_ratio': 1},
-            phi_hyp_prior={'mean': 3, 'sd': 1}
+            phi_hyp_prior={'mean': 3, 'sd': 1},
+            spot_fact_mean_var_ratio=5,
+            exper_gene_level_mean_var_ratio=10,
     ):
+
         ############# Initialise parameters ################
         super().__init__(cell_state_mat, X_data,
                          data_type, n_iter,
@@ -71,18 +136,31 @@ class LocationModelHierarchicalW(Pymc3LocModel):
 
         for k in gene_level_var_prior.keys():
             gene_level_prior[k] = gene_level_var_prior[k]
-        self.gene_level_prior = gene_level_prior
 
+        self.gene_level_prior = gene_level_prior
+        self.phi_hyp_prior = phi_hyp_prior
+        self.n_comb = n_comb
+        self.spot_fact_mean_var_ratio = spot_fact_mean_var_ratio
+        self.exper_gene_level_mean_var_ratio = exper_gene_level_mean_var_ratio
+        
+        # generate parameters for samples
+        self.spot2sample_df = pd.get_dummies(sample_id)
+        # convert to np.ndarray
+        self.spot2sample_mat = self.spot2sample_df.values
+        self.n_exper = self.spot2sample_mat.shape[1]
+        # assign extra data to dictionary with (1) shared parameters (2) input data
+        self.extra_data_tt = {'spot2sample': theano.shared(self.spot2sample_mat.astype(self.data_type))}
+        self.extra_data = {'spot2sample': self.spot2sample_mat.astype(self.data_type)}
+        
         for k in cell_number_var_prior.keys():
             cell_number_prior[k] = cell_number_var_prior[k]
         self.cell_number_prior = cell_number_prior
-
-        self.phi_hyp_prior = phi_hyp_prior
 
         ############# Define the model ################
         self.model = pm.Model()
 
         with self.model:
+
             # =====================Gene expression level scaling======================= #
             # Explains difference in expression between genes and 
             # how it differs in single cell and spatial technology
@@ -97,12 +175,13 @@ class LocationModelHierarchicalW(Pymc3LocModel):
             self.gene_level_beta_hyp = pm.Gamma('gene_level_beta_hyp',
                                                 mu=rate, sigma=np.sqrt(rate_var),
                                                 shape=(1, 1))
-
+            
+            # global gene levels
             self.gene_level = pm.Gamma('gene_level', self.gene_level_alpha_hyp,
                                        self.gene_level_beta_hyp, shape=(self.n_var, 1))
-
             # scale cell state factors by gene_level
             self.gene_factors = pm.Deterministic('gene_factors', self.cell_state)
+            #self.gene_factors = self.cell_state
             # tt.printing.Print('gene_factors sum')(gene_factors.sum(0).shape)
             # tt.printing.Print('gene_factors sum')(gene_factors.sum(0))
 
@@ -128,7 +207,7 @@ class LocationModelHierarchicalW(Pymc3LocModel):
             # =====================Spot-specific additive component======================= #
             # molecule contribution that cannot be explained by cell state signatures
             # these counts are distributed between all genes not just expressed genes
-            self.spot_add_hyp = pm.Gamma('spot_add_hyp', 1, 0.1, shape=2)
+            self.spot_add_hyp = pm.Gamma('spot_add_hyp', 1, 1, shape=2)
             self.spot_add = pm.Gamma('spot_add', self.spot_add_hyp[0],
                                      self.spot_add_hyp[1], shape=(self.n_obs, 1))
 
@@ -137,23 +216,25 @@ class LocationModelHierarchicalW(Pymc3LocModel):
             # these counts are distributed equally between all spots (e.g. background, free-floating RNA)
             self.gene_add_hyp = pm.Gamma('gene_add_hyp', 1, 1, shape=2)
             self.gene_add = pm.Gamma('gene_add', self.gene_add_hyp[0],
-                                     self.gene_add_hyp[1], shape=(self.n_var, 1))
+                                     self.gene_add_hyp[1], shape=(self.n_exper, self.n_var))
 
             # =====================Gene-specific overdispersion ======================= #
             self.phi_hyp = pm.Gamma('phi_hyp', mu=phi_hyp_prior['mean'],
                                     sigma=phi_hyp_prior['sd'], shape=(1, 1))
-            self.gene_E = pm.Exponential('gene_E', self.phi_hyp, shape=(self.n_var, 1))
+            self.gene_E = pm.Exponential('gene_E', self.phi_hyp, shape=(self.n_exper, self.n_var))
 
             # =====================Expected expression ======================= #
             # expected expression
-            self.mu_biol = pm.math.dot(self.spot_factors, self.gene_factors.T) * self.gene_level.T \
-                           + self.gene_add.T + self.spot_add
+            self.mu_biol = pm.math.dot(self.spot_factors, self.gene_factors.T) \
+                           * self.gene_level.T \
+                           + pm.math.dot(self.extra_data_tt['spot2sample'], self.gene_add) + self.spot_add
             # tt.printing.Print('mu_biol')(self.mu_biol.shape)
 
             # =====================DATA likelihood ======================= #
             # Likelihood (sampling distribution) of observations & add overdispersion via NegativeBinomial / Poisson
             self.data_target = pm.NegativeBinomial('data_target', mu=self.mu_biol,
-                                                   alpha=1 / (self.gene_E.T * self.gene_E.T),
+                                                   alpha=pm.math.dot(self.extra_data_tt['spot2sample'], 
+                                                                     1 / tt.pow(self.gene_E, 2)),
                                                    observed=self.x_data,
                                                    total_size=self.X_data.shape)
 
@@ -163,14 +244,15 @@ class LocationModelHierarchicalW(Pymc3LocModel):
 
     def compute_expected(self):
         r"""Compute expected expression of each gene in each spot (Poisson mu). Useful for evaluating how well
-        the model learned expression pattern of all genes in the data.
-
+            the model learned expression pattern of all genes in the data.
         """
 
         # compute the poisson rate
         self.mu = (np.dot(self.samples['post_sample_means']['spot_factors'],
                           self.samples['post_sample_means']['gene_factors'].T)
                    * self.samples['post_sample_means']['gene_level'].T
-                   + self.samples['post_sample_means']['gene_add'].T
+                   + np.dot(self.extra_data['spot2sample'],
+                            self.samples['post_sample_means']['gene_add'])
                    + self.samples['post_sample_means']['spot_add'])
-
+        self.alpha = np.dot(self.extra_data['spot2sample'],
+                            1 / (self.samples['post_sample_means']['gene_E'] * self.samples['post_sample_means']['gene_E']))
