+++ cell2location/models/simplified/LocationModelLinearDependentWMultiExperimentNoSegLs.py
@@ -1,39 +1,40 @@
 # -*- coding: utf-8 -*-
-r"""The Co-Location model decomposes the expression of genes across locations into a set
-    of reference regulatory programmes, while accounting for correlation of programs
-    across locations with similar cell composition."""
+r"""The Co-Location model decomposes the expression of genes across locations measured in multiple experiments
+    into a set of reference regulatory programmes, while accounting for correlation of programs
+    across locations with similar cell composition.
+    Overdispersion alpha_eg & additive background s_eg for each experiment and gene."""
 
 import numpy as np
+import pandas as pd
 import pymc3 as pm
 import theano.tensor as tt
+import theano
 
-from cell2location.models.pymc3_loc_model import Pymc3LocModel
+from cell2location.models.base.pymc3_loc_model import Pymc3LocModel
 
 
 # defining the model itself
-class LocationModelLinearDependentW(Pymc3LocModel):
+class LocationModelLinearDependentWMultiExperimentNoSegLs(Pymc3LocModel):
     r"""Cell2location models the elements of :math:`D` as Negative Binomial distributed,
     given an unobserved rate :math:`mu` and a gene-specific over-dispersion parameter :math:`\alpha_g`
     which describes variance in expression of individual genes that is not explained by the regulatory programs:
     
     .. math::
-        D_{s,g} \sim \mathtt{NB}(\mu_{s,g}, \alpha_g)
+        D_{s,g} \sim \mathtt{NB}(\mu_{s,g}, \alpha_{e,g})
     
-    The containment prior on overdispersion :math:`\alpha_g` parameter is used
+    The containment prior on overdispersion :math:`\alpha_{e,g}` parameter is used
     (for more details see: https://statmodeling.stat.columbia.edu/2018/04/03/justify-my-love/).
     
     The spatial expression levels of genes :math:`\mu_{s,g}` in the rate space are modelled
     as the sum of five non-negative components:
     
     .. math::
-        \mu_{s,g} = m_{g} \left (\sum_{f} {w_{s,f} \: g_{f,g}} \right) + l_s + s_{g}
+        \mu_{s,g} = m_{g} \left (\sum_{f} {w_{s,f} \: g_{f,g}} \right) 
     
     Here, :math:`w_{s,f}` denotes regression weight of each program :math:`f` at location :math:`s` ;
     :math:`g_{f,g}` denotes the reference signatures of cell types :math:`f` of each gene :math:`g` - input to the model;
     :math:`m_{g}` denotes a gene-specific scaling parameter which accounts for difference
     in the global expression estimates between technologies;
-    :math:`l_{s}` and :math:`s_{g}` are additive components that capture additive background variation
-    that is not explained by the bi-variate decomposition.
     
     The prior distribution on :math:`w_{s,f}` is chosen to reflect the absolute scale and account for correlation of programs
     across locations with similar cell composition. This is done by inferring a hierarchical prior representing
@@ -119,7 +120,7 @@ class LocationModelLinearDependentW(Pymc3LocModel):
     -------
 
     """
-# CoLocationModelNB4V2
+# CoLocationModelNB4E6V2
     def __init__(
             self,
             cell_state_mat: np.ndarray,
@@ -141,7 +142,8 @@ class LocationModelLinearDependentW(Pymc3LocModel):
                                    'factors_mean_var_ratio': 1,
                                    'combs_mean_var_ratio': 1},
             phi_hyp_prior={'mean': 3, 'sd': 1},
-            spot_fact_mean_var_ratio=5
+            spot_fact_mean_var_ratio=5,
+            exper_gene_level_mean_var_ratio=10,
     ):
 
         ############# Initialise parameters ################
@@ -158,6 +160,16 @@ class LocationModelLinearDependentW(Pymc3LocModel):
         self.phi_hyp_prior = phi_hyp_prior
         self.n_comb = n_comb
         self.spot_fact_mean_var_ratio = spot_fact_mean_var_ratio
+        self.exper_gene_level_mean_var_ratio = exper_gene_level_mean_var_ratio
+        
+        # generate parameters for samples
+        self.spot2sample_df = pd.get_dummies(sample_id)
+        # convert to np.ndarray
+        self.spot2sample_mat = self.spot2sample_df.values
+        self.n_exper = self.spot2sample_mat.shape[1]
+        # assign extra data to dictionary with (1) shared parameters (2) input data
+        self.extra_data_tt = {'spot2sample': theano.shared(self.spot2sample_mat.astype(self.data_type))}
+        self.extra_data = {'spot2sample': self.spot2sample_mat.astype(self.data_type)}
 
         cell_number_prior['factors_per_combs'] = (cell_number_prior['factors_per_spot'] /
                                                   cell_number_prior['combs_per_spot'])
@@ -178,23 +190,19 @@ class LocationModelLinearDependentW(Pymc3LocModel):
             rate = gene_level_prior['mean'] / gene_level_prior['sd'] ** 2
             shape_var = shape / gene_level_prior['mean_var_ratio']
             rate_var = rate / gene_level_prior['mean_var_ratio']
-            n_g_prior = np.array(gene_level_prior['mean']).shape
-            if len(n_g_prior) == 0:
-                n_g_prior = 1
-            else:
-                n_g_prior = self.n_var
             self.gene_level_alpha_hyp = pm.Gamma('gene_level_alpha_hyp',
                                                  mu=shape, sigma=np.sqrt(shape_var),
-                                                    shape=(n_g_prior, 1))
+                                                 shape=(1, 1))
             self.gene_level_beta_hyp = pm.Gamma('gene_level_beta_hyp',
                                                 mu=rate, sigma=np.sqrt(rate_var),
-                                                shape=(n_g_prior, 1))
-
+                                                shape=(1, 1))
+            
+            # global gene levels
             self.gene_level = pm.Gamma('gene_level', self.gene_level_alpha_hyp,
                                        self.gene_level_beta_hyp, shape=(self.n_var, 1))
-
             # scale cell state factors by gene_level
             self.gene_factors = pm.Deterministic('gene_factors', self.cell_state)
+            #self.gene_factors = self.cell_state
             # tt.printing.Print('gene_factors sum')(gene_factors.sum(0).shape)
             # tt.printing.Print('gene_factors sum')(gene_factors.sum(0))
 
@@ -222,7 +230,6 @@ class LocationModelLinearDependentW(Pymc3LocModel):
                                               sigma=np.sqrt(cell_number_prior['factors_per_combs'] \
                                                             / cell_number_prior['factors_mean_var_ratio']),
                                               shape=(self.n_comb, 1))
-
             c2f_shape = self.factors_per_combs / np.array(self.n_fact).reshape((1, 1))
             self.comb2fact = pm.Gamma('comb2fact', alpha=c2f_shape, beta=self.factors_per_combs,
                                       shape=(self.n_comb, self.n_fact))
@@ -232,35 +239,22 @@ class LocationModelLinearDependentW(Pymc3LocModel):
                                                             / self.spot_fact_mean_var_ratio),
                                          shape=(self.n_obs, self.n_fact))
 
-            # =====================Spot-specific additive component======================= #
-            # molecule contribution that cannot be explained by cell state signatures
-            # these counts are distributed between all genes not just expressed genes
-            self.spot_add_hyp = pm.Gamma('spot_add_hyp', 1, 1, shape=2)
-            self.spot_add = pm.Gamma('spot_add', self.spot_add_hyp[0],
-                                     self.spot_add_hyp[1], shape=(self.n_obs, 1))
-
-            # =====================Gene-specific additive component ======================= #
-            # per gene molecule contribution that cannot be explained by cell state signatures
-            # these counts are distributed equally between all spots (e.g. background, free-floating RNA)
-            self.gene_add_hyp = pm.Gamma('gene_add_hyp', 1, 1, shape=2)
-            self.gene_add = pm.Gamma('gene_add', self.gene_add_hyp[0],
-                                     self.gene_add_hyp[1], shape=(self.n_var, 1))
-
             # =====================Gene-specific overdispersion ======================= #
             self.phi_hyp = pm.Gamma('phi_hyp', mu=phi_hyp_prior['mean'],
                                     sigma=phi_hyp_prior['sd'], shape=(1, 1))
-            self.gene_E = pm.Exponential('gene_E', self.phi_hyp, shape=(self.n_var, 1))
+            self.gene_E = pm.Exponential('gene_E', self.phi_hyp, shape=(self.n_exper, self.n_var))
 
             # =====================Expected expression ======================= #
             # expected expression
-            self.mu_biol = pm.math.dot(self.spot_factors, self.gene_factors.T) * self.gene_level.T \
-                           + self.gene_add.T + self.spot_add
+            self.mu_biol = pm.math.dot(self.spot_factors, self.gene_factors.T) \
+                           * self.gene_level.T 
             # tt.printing.Print('mu_biol')(self.mu_biol.shape)
 
             # =====================DATA likelihood ======================= #
             # Likelihood (sampling distribution) of observations & add overdispersion via NegativeBinomial / Poisson
             self.data_target = pm.NegativeBinomial('data_target', mu=self.mu_biol,
-                                                   alpha=1 / (self.gene_E.T * self.gene_E.T),
+                                                   alpha=pm.math.dot(self.extra_data_tt['spot2sample'], 
+                                                                     1 / tt.pow(self.gene_E, 2)),
                                                    observed=self.x_data,
                                                    total_size=self.X_data.shape)
 
@@ -276,7 +270,6 @@ class LocationModelLinearDependentW(Pymc3LocModel):
         # compute the poisson rate
         self.mu = (np.dot(self.samples['post_sample_means']['spot_factors'],
                           self.samples['post_sample_means']['gene_factors'].T)
-                   * self.samples['post_sample_means']['gene_level'].T
-                   + self.samples['post_sample_means']['gene_add'].T
-                   + self.samples['post_sample_means']['spot_add'])
-        self.alpha = 1 / (self.samples['post_sample_means']['gene_E'].T * self.samples['post_sample_means']['gene_E'].T)
+                   * self.samples['post_sample_means']['gene_level'].T)
+        self.alpha = np.dot(self.extra_data['spot2sample'],
+                            1 / (self.samples['post_sample_means']['gene_E'] * self.samples['post_sample_means']['gene_E']))
